+++ pkg/ingester-rf1/flush.go
@@ -1,0 +1,159 @@
+package ingesterrf1
+
+import (
+	"crypto/rand"
+	"fmt"
+	"net/http"
+	"strconv"
+	"time"
+
+	"github.com/dustin/go-humanize"
+	"github.com/go-kit/log"
+	"github.com/go-kit/log/level"
+	"github.com/grafana/dskit/backoff"
+	"github.com/grafana/dskit/ring"
+	"github.com/grafana/dskit/runutil"
+	"github.com/oklog/ulid"
+	"golang.org/x/net/context"
+
+	"github.com/grafana/loki/v3/pkg/storage/wal"
+	"github.com/grafana/loki/v3/pkg/util"
+	util_log "github.com/grafana/loki/v3/pkg/util/log"
+)
+
+const (
+	// Backoff for retrying 'immediate' flushes. Only counts for queue
+	// position, not wallclock time.
+	flushBackoff = 1 * time.Second
+
+	nameLabel = "__name__"
+	logsValue = "logs"
+
+	flushReasonIdle   = "idle"
+	flushReasonMaxAge = "max_age"
+	flushReasonForced = "forced"
+	flushReasonFull   = "full"
+	flushReasonSynced = "synced"
+)
+
+// Note: this is called both during the WAL replay (zero or more times)
+// and then after replay as well.
+func (i *Ingester) InitFlushQueues() {
+	i.flushQueuesDone.Add(i.cfg.ConcurrentFlushes)
+	for j := 0; j < i.cfg.ConcurrentFlushes; j++ {
+		i.flushQueues[j] = util.NewPriorityQueue(i.metrics.flushQueueLength)
+		go i.flushLoop(j)
+	}
+}
+
+// Flush implements ring.FlushTransferer
+// Flush triggers a flush of all the chunks and closes the flush queues.
+// Called from the Lifecycler as part of the ingester shutdown.
+func (i *Ingester) Flush() {
+	i.flush()
+}
+
+// TransferOut implements ring.FlushTransferer
+// Noop implemenetation because ingesters have a WAL now that does not require transferring chunks any more.
+// We return ErrTransferDisabled to indicate that we don't support transfers, and therefore we may flush on shutdown if configured to do so.
+func (i *Ingester) TransferOut(_ context.Context) error {
+	return ring.ErrTransferDisabled
+}
+
+func (i *Ingester) flush() {
+	// TODO: Flush the last chunks
+	// Close the flush queues, to unblock waiting workers.
+	for _, flushQueue := range i.flushQueues {
+		flushQueue.Close()
+	}
+
+	i.flushQueuesDone.Wait()
+	level.Debug(i.logger).Log("msg", "flush queues have drained")
+}
+
+// FlushHandler triggers a flush of all in memory chunks.  Mainly used for
+// local testing.
+func (i *Ingester) FlushHandler(w http.ResponseWriter, _ *http.Request) {
+	w.WriteHeader(http.StatusNoContent)
+}
+
+type flushOp struct {
+	it  *wal.PendingItem
+	num int64
+}
+
+func (o *flushOp) Key() string {
+	return strconv.Itoa(int(o.num))
+}
+
+func (o *flushOp) Priority() int64 {
+	return -o.num
+}
+
+func (i *Ingester) flushLoop(j int) {
+	l := log.With(i.logger, "loop", j)
+	defer func() {
+		level.Debug(l).Log("msg", "Ingester.flushLoop() exited")
+		i.flushQueuesDone.Done()
+	}()
+
+	for {
+		o := i.flushQueues[j].Dequeue()
+		if o == nil {
+			return
+		}
+		op := o.(*flushOp)
+
+		start := time.Now()
+
+		// We'll use this to log the size of the segment that was flushed.
+		n := humanize.Bytes(uint64(op.it.Writer.InputSize()))
+
+		err := i.flushOp(l, op)
+		d := time.Since(start)
+		if err != nil {
+			level.Error(l).Log("msg", "failed to flush", "size", n, "duration", d, "err", err)
+		} else {
+			level.Debug(l).Log("msg", "flushed", "size", n, "duration", d)
+		}
+
+		op.it.Result.SetDone(err)
+		i.wal.Put(op.it)
+	}
+}
+
+func (i *Ingester) flushOp(l log.Logger, op *flushOp) error {
+	ctx, cancelFunc := context.WithCancel(context.Background())
+	defer cancelFunc()
+
+	b := backoff.New(ctx, i.cfg.FlushOpBackoff)
+	for b.Ongoing() {
+		err := i.flushSegment(ctx, op.it.Writer)
+		if err == nil {
+			break
+		}
+		level.Error(l).Log("msg", "failed to flush", "retries", b.NumRetries(), "err", err)
+		b.Wait()
+	}
+	return b.Err()
+}
+
+// flushChunk flushes the given chunk to the store.
+//
+// If the flush is successful, metrics for this flush are to be reported.
+// If the flush isn't successful, the operation for this userID is requeued allowing this and all other unflushed
+// segments to have another opportunity to be flushed.
+func (i *Ingester) flushSegment(ctx context.Context, ch *wal.SegmentWriter) error {
+	reader := ch.Reader()
+	defer runutil.CloseWithLogOnErr(util_log.Logger, reader, "flushSegment")
+
+	newUlid := ulid.MustNew(ulid.Timestamp(time.Now()), rand.Reader)
+
+	if err := i.store.PutObject(ctx, fmt.Sprintf("loki-v2/wal/anon/"+newUlid.String()), reader); err != nil {
+		i.metrics.chunksFlushFailures.Inc()
+		return fmt.Errorf("store put chunk: %w", err)
+	}
+	i.metrics.flushedChunksStats.Inc(1)
+	// TODO: report some flush metrics
+	return nil
+}